+++ code/agent/common/util.py
@@ -15,11 +15,11 @@ from subprocess import (Popen, run, PIPE, TimeoutExpired,
 
 base_label = 'nuvlaedge.component=True'
 compose_project_name = os.getenv('COMPOSE_PROJECT_NAME', 'nuvlaedge')
-compute_api_port = os.getenv('COMPUTE_API_PORT', '5000')
 default_project_name = 'nuvlaedge'
 
 
 COMPUTE_API_INTERNAL_PORT = 5000
+
 
 def str_if_value_or_none(value):
     return str(value) if value else None