+++ code/tests/monitor/components/test_container_stats.py
@@ -211,6 +211,11 @@ class TestContainerStatsMonitor(unittest.TestCase):
         test_monitor.populate_nb_report(nb_report)
         self.assertIn('some_Data', nb_report)
 
+        container_data = Mock()
+        container_data.name = 'container_name'
+        test_monitor.data.containers = {'container': container_data}
+        test_monitor.populate_nb_report(nb_report)
+
         test_monitor: ContainerStatsMonitor = self.get_base_monitor()
         test_monitor.data = Mock()
         test_monitor.data.cluster_data = None